requirements.txt
opencv-python
pyzbar
Pillow

